feat(KinoTermite.Adapter): xtermjs adapter for LiveBook
The KinoTermite adapter allows for using Termite within a LiveBook
environment using KinoLive and xterm.js. The `convertEol` has been set
for xterm which converts "\n" to "\r\n" - this maintains compatability
with the examples from the termite library.

The `addon-fit` is used, which handles resize events from the terminal.
Since we need to get the size from the browser, on first load, we set it
to a known value, however after any output has been written to the
terminal, any future resize events will be requested from the browser.